+++ src/pages/channel/ui/chat-item/chat-item.tsx
