+++ src/pages/channel/ui/chat-item/constants.ts
@@ -1,0 +1,1 @@
+export const CHAT_PAGE = 'chats';
