+++ src/pages/channel/ui/chat-item/index.ts
@@ -1,0 +1,1 @@
+export { ChatItem } from './chat-item';
