+++ src/pages/channel/ui/chat-item/types.ts
@@ -1,0 +1,6 @@
+import type { ChatModel } from '../../model';
+
+export type ChatItemProps = {
+  chat: ChatModel;
+  isActive: boolean;
+};
