+++ src/pages/channel/ui/chat-list/index.ts
@@ -1,0 +1,2 @@
+export { ChatList } from './chat-list';
+export type { ChatListProps } from './types';
