+++ src/pages/channel/ui/chat-list/types.ts
@@ -1,0 +1,3 @@
+export type ChatListProps = {
+  chatId: string;
+};
