+++ src/widgets/app-left-sidebar/model/channel.model.ts
