--- src/widgets/app-left-sidebar/ui/app-left-sidebar/channels/components/index.ts
@@ -1,2 +1,0 @@
-export { CHANNEL_PAGE } from './channel-avatar';
-export { ChannelList } from './channel-list';
