--- src/widgets/app-left-sidebar/ui/app-left-sidebar/channels/index.ts
@@ -1,2 +1,0 @@
-// NOTE: залог на entity
-export { CHANNEL_PAGE, ChannelList } from './components';
