+++ src/widgets/app-left-sidebar/ui/app-left-sidebar/index.ts
@@ -1,2 +1,1 @@
 export { AppLeftSidebar } from './app-left-sidebar';
-export { CHANNEL_PAGE } from './channels';
