+++ src/widgets/app-left-sidebar/ui/channel-avatar/channel-avatar.tsx
