+++ src/widgets/app-left-sidebar/ui/channel-avatar/constants.ts
