+++ src/widgets/app-left-sidebar/ui/channel-avatar/index.ts
